Play button update
Now it is responsible for playing as well as reseting the song. Its name got changed to reflect its new funciton. It will be activated only when a correct directory will be chosen, before the rest of the buttons. Now it has two stages: on the first click it plays the song,  the second click resets the song and so on it that order. Every if and else concerning other buttons got removed since they are no longer necessary