Program update
Loop current track button has been added. It has two states an active state and sleeping state. When active it loops a track and when sleeping it does nothing. Both forward and rewind track buttons got a rework and also they properly work with the new loop button, now. Reset button resets a track when it ends and when the loop button is not active. And other major fixes and changes were made (more cosmetic ones I think)